ehp-core-main/requirements/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/requirements.in -o requirements/requirements.txt
aiofiles==24.1.0
    # via -r requirements/requirements.in
alembic==1.15.2
    # via -r requirements/requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
asyncio==3.4.3
    # via -r requirements/requirements.in
asyncpg==0.30.0
    # via -r requirements/requirements.in
autoflake==2.3.1
    # via -r requirements/requirements.in
black==25.1.0
    # via -r requirements/requirements.in
boto3==1.38.25
    # via
    #   -r requirements/requirements.in
    #   moto
boto3-stubs==1.38.26
    # via -r requirements/requirements.in
botocore==1.38.25
    # via
    #   boto3
    #   moto
    #   s3transfer
botocore-stubs==1.38.26
    # via boto3-stubs
certifi==2025.4.26
    # via
    #   elastic-transport
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via
    #   black
    #   uvicorn
coverage==7.8.0
    # via
    #   -r requirements/requirements.in
    #   pytest-cov
cryptography==45.0.3
    # via moto
cuid==0.4
    # via -r requirements/requirements.in
elastic-transport==8.17.1
    # via elasticsearch
elasticsearch==9.0.1
    # via -r requirements/requirements.in
fastapi==0.115.12
    # via
    #   -r requirements/requirements.in
    #   fastapi-sessions
fastapi-sessions==0.3.2
    # via -r requirements/requirements.in
flake8==7.2.0
    # via -r requirements/requirements.in
flake8-import-order==0.18.2
    # via -r requirements/requirements.in
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via -r requirements/requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.1.0
    # via pytest
invoke==2.2.0
    # via -r requirements/requirements.in
isort==6.0.1
    # via -r requirements/requirements.in
itsdangerous==2.2.0
    # via fastapi-sessions
jinja2==3.1.6
    # via
    #   -r requirements/requirements.in
    #   moto
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
linecache2==1.0.0
    # via traceback2
mako==1.3.10
    # via alembic
markupsafe==3.0.2
    # via
    #   jinja2
    #   mako
    #   werkzeug
mccabe==0.7.0
    # via flake8
moto==5.1.5
    # via -r requirements/requirements.in
mypy==1.15.0
    # via -r requirements/requirements.in
mypy-boto3-secretsmanager==1.38.0
    # via boto3-stubs
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
packaging==25.0
    # via
    #   black
    #   pytest
pathspec==0.12.1
    # via black
platformdirs==4.3.8
    # via black
pluggy==1.5.0
    # via pytest
psycopg2-binary==2.9.10
    # via -r requirements/requirements.in
pycodestyle==2.13.0
    # via
    #   flake8
    #   flake8-import-order
pycparser==2.22
    # via cffi
pycryptodome==3.22.0
    # via -r requirements/requirements.in
pydantic==2.11.4
    # via
    #   -r requirements/requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.9.1
    # via -r requirements/requirements.in
pyflakes==3.3.2
    # via
    #   autoflake
    #   flake8
pyjwt==2.10.1
    # via -r requirements/requirements.in
pytest==8.3.5
    # via
    #   -r requirements/requirements.in
    #   pytest-asyncio
    #   pytest-cov
    #   pytest-mock
pytest-asyncio==0.26.0
    # via -r requirements/requirements.in
pytest-cov==6.1.1
    # via -r requirements/requirements.in
pytest-mock==3.14.0
    # via -r requirements/requirements.in
python-dateutil==2.9.0.post0
    # via
    #   -r requirements/requirements.in
    #   botocore
    #   elasticsearch
    #   moto
python-dotenv==1.1.0
    # via
    #   -r requirements/requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements/requirements.in
pyyaml==6.0.2
    # via
    #   responses
    #   uvicorn
redis==6.0.0
    # via -r requirements/requirements.in
requests==2.32.3
    # via
    #   -r requirements/requirements.in
    #   moto
    #   responses
responses==0.25.7
    # via moto
s3transfer==0.13.0
    # via boto3
setuptools==80.9.0
    # via flake8-import-order
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.30
    # via
    #   -r requirements/requirements.in
    #   alembic
starlette==0.46.2
    # via fastapi
traceback2==1.4.0
    # via -r requirements/requirements.in
types-awscrt==0.27.2
    # via botocore-stubs
types-s3transfer==0.13.0
    # via boto3-stubs
typing-extensions==4.13.2
    # via
    #   alembic
    #   anyio
    #   boto3-stubs
    #   elasticsearch
    #   fastapi
    #   mypy
    #   mypy-boto3-secretsmanager
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.0
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.4.0
    # via
    #   botocore
    #   elastic-transport
    #   requests
    #   responses
uuid==1.30
    # via -r requirements/requirements.in
uvicorn==0.34.2
    # via -r requirements/requirements.in
uvloop==0.21.0
    # via uvicorn
vulture==2.14
    # via -r requirements/requirements.in
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn
werkzeug==3.1.3
    # via
    #   -r requirements/requirements.in
    #   moto
xmltodict==0.14.2
    # via moto
